- Change mb_copy_t to take a size_t as the length argument instead of an   int. - Change the local variable in smb_copy_iconv() from an int to a size_t.
These make smb_copy_iconv() happy in a 64-bit world.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@106667 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/netsmb/smb_subr.c
@@ -307,9 +307,9 @@ smb_maperror(int eclass, int eno)
 }
 
 static int
-smb_copy_iconv(struct mbchain *mbp, c_caddr_t src, caddr_t dst, int len)
+smb_copy_iconv(struct mbchain *mbp, c_caddr_t src, caddr_t dst, size_t len)
 {
-	int outlen = len;
+	size_t outlen = len;
 
 	return iconv_conv((struct iconv_drv*)mbp->mb_udata, &src, &len, &dst, &outlen);
 }
